Remove all the logic below the application-layer
MQTT is defined on top of the applicaiton-layer and all that it requires is an
active session. We can simplify the logic by assuming a user has performed
connection to a broker in advance and supplies the session.

It is nice and kind of the `rumqtt` client implementation where it takes care of
all the TLS connection building, like CA and certs. However it is only possible
because of `tokio_tls` and not the case for embedded environment.