Propose Session trait
Although it look pedantic I propose a session trait encapsulating network
transaction in general. One downside of the `TcpClient` trait is that it only
allows a single socket type. On the other hand, the `Session` trait can be
implemented by multiple transaction types and they can share the same driver
type of `TcpClient`.